feat: Add sandbox and hls media server
So, I do not know rust very well, this is a learning project.
I've never built a media player and learning how streaming music
works is facinating. The idea is for the media server to well...
serve music. That will likely be stored in AWS because GCP pricing
is absolute madness in comparison. Plan is for there to eventually
be a rust layer for the frontend that we can then plug into multiple
different UI's.